% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicalVariantR.R
\name{ClinicalVariantR}
\alias{ClinicalVariantR}
\title{Build the ClinicalVariantR Shiny application}
\usage{
ClinicalVariantR(...)
}
\arguments{
\item{...}{Reserved for future options (currently ignored).}
}
\value{
A \code{shiny.appobj} created by \code{shiny::shinyApp()}.
}
\description{
Returns a Shiny app object for ACMG/AMP germline variant classification.
Following Bioconductor Shiny guidelines, this function \strong{returns} the app
and does not launch it. Call \code{shiny::runApp()} yourself.
}
\examples{
stopifnot(is.function(ClinicalVariantR))
stopifnot(is.function(ClinicalVariantRApp))

}
\seealso{
\code{\link{ClinicalVariantRApp}}
}
