% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicalVariantR.R
\name{ClinicalVariantRApp}
\alias{ClinicalVariantRApp}
\title{Alias for \code{\link{ClinicalVariantR}}}
\usage{
ClinicalVariantRApp(...)
}
\arguments{
\item{...}{Reserved for future options (currently ignored).}
}
\value{
A \code{shiny.appobj} created by \code{shiny::shinyApp()}.
}
\description{
Alias for \code{\link{ClinicalVariantR}}
}
\examples{
stopifnot(is.function(ClinicalVariantRApp))
}
